+++ build.js
@@ -6,8 +6,8 @@ var fable = require("fable-compiler");
 var targets = {
     all() {
         return fable.promisify(fs.remove, "npm")
-            .then(_ => fable.compile({ verbose: true }))
-            // .then(_ => fable.compile({target: "umd"}))
+            .then(_ => fable.compile())
+            .then(_ => fable.compile({target: "umd"}))
             .then(_ => fable.promisify(fs.copy, "package.json", "npm/package.json"))
             .then(_ => fable.promisify(fs.copy, "README.md", "npm/README.md"))
             .then(_ => fable.promisify(fs.readFile, "RELEASE_NOTES.md"))